service, repository 1차 완료 -> controller 시작

+++ foobar/src/main/java/team/foobar/service/comment/CommentServiceImpl.java
@@ -3,8 +3,11 @@ package team.foobar.service.comment;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Primary;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.StringUtils;
 import team.foobar.domain.Board;
 import team.foobar.domain.Comment;
 import team.foobar.domain.Member;
@@ -13,12 +16,13 @@ import team.foobar.repository.jpa.board.BoardRepository;
 import team.foobar.repository.jpa.comment.CommentRepository;
 import team.foobar.repository.jpa.member.MemberRepository;
 
+import javax.print.attribute.standard.PresentationDirection;
 import java.util.List;
 import java.util.Optional;
 
 @Slf4j
 @Service
-@Transactional
+@Transactional(readOnly = true)
 @RequiredArgsConstructor
 @Primary
 public class CommentServiceImpl implements CommentService {
@@ -29,6 +33,16 @@ public class CommentServiceImpl implements CommentService {
     @Override
     public List<Comment> searchByBoardId(Integer id) {
         return repository.findByBoardId(id);
+    }
+
+    @Override
+    public Optional<Comment> search(Integer id) {
+        return repository.findByIdWithFetch(id);
+    }
+
+    @Override
+    public Page<Comment> searchByMemberId(Integer id, Pageable pageable) {
+        return repository.findByMemberIdWithFetch(id, pageable);
     }
 
     @Override
@@ -45,6 +59,7 @@ public class CommentServiceImpl implements CommentService {
         }
 
         Comment save = repository.save(dtoToEntity(dto));
+        log.info("save = {}", save);
         return Optional.of(save.getId());
     }
 
@@ -59,7 +74,8 @@ public class CommentServiceImpl implements CommentService {
         Comment comment = search.get();
         comment.change(
                 dto.getTagMemberId() != null ? Member.builder().id(dto.getTagMemberId()).build() : null,
-                dto.getContent()
+                dto.getContent(),
+                null
         );
 
         return Optional.of(comment.getId());
@@ -68,6 +84,24 @@ public class CommentServiceImpl implements CommentService {
     @Override
     @Transactional
     public void delete(Integer id) {
-        repository.deleteById(id);
+        Optional<Comment> search = repository.findByIdWithFetch(id);
+        if(search.isEmpty()) {
+            return;
+        }
+
+        /* 자식댓글이면 삭제 */
+        /* 부모댓글이면 자식있으면 deleteFl 변경, 없으면 그냥 삭제 */
+        Comment comment = search.get();
+        if(comment.getParent() != null) {
+            repository.deleteById(comment.getId());
+        } else {
+            List<Comment> childList = repository.findChildById(id);
+
+            if(childList.size() == 0) {
+                repository.deleteById(id);
+            } else {
+                comment.change(null, null, true);
+            }
+        }
     }
 }
